chore(packaging): move cstor related packages within cstor package
This commit tries to reduce the number of packages that has
increased recently due to adoption of builder pattern. In this
commit following cstor packages are moved to pkg/cstor:
- pool,
- poolcluster,
- volume,
- volumereplica,
- grpc server & client
