+++ pkg/algorithm/nodeselect/v1alpha1/config.go
@@ -18,10 +18,9 @@ package v1alpha1
 
 import (
 	apis "github.com/openebs/maya/pkg/apis/openebs.io/v1alpha1"
-	"github.com/openebs/maya/pkg/client/k8s"
-	cstorpool "github.com/openebs/maya/pkg/cstorpool/v1alpha1"
-
 	blockdevice "github.com/openebs/maya/pkg/blockdevice/v1alpha1"
+	"github.com/openebs/maya/pkg/client/k8s"
+	cstorpool "github.com/openebs/maya/pkg/cstor/pool/v1alpha1"
 	env "github.com/openebs/maya/pkg/env/v1alpha1"
 	sp "github.com/openebs/maya/pkg/sp/v1alpha1"
 )
