+++ pkg/algorithm/nodeselect/v1alpha1/select_node_test.go
@@ -17,12 +17,8 @@ limitations under the License.
 package v1alpha1
 
 import (
-	"testing"
-
-	"k8s.io/client-go/kubernetes/fake"
-
-	//openebsFakeClientset "github.com/openebs/maya/pkg/client/clientset/versioned/fake"
 	"strconv"
+	"testing"
 
 	"github.com/golang/glog"
 	ndmapis "github.com/openebs/maya/pkg/apis/openebs.io/ndm/v1alpha1"
@@ -30,9 +26,10 @@ import (
 	blockdevice "github.com/openebs/maya/pkg/blockdevice/v1alpha1"
 	openebsFakeClientset "github.com/openebs/maya/pkg/client/generated/clientset/versioned/fake"
 	ndmFakeClientset "github.com/openebs/maya/pkg/client/generated/openebs.io/ndm/v1alpha1/clientset/internalclientset/fake"
-	cstorpool "github.com/openebs/maya/pkg/cstorpool/v1alpha1"
+	cstorpool "github.com/openebs/maya/pkg/cstor/pool/v1alpha1"
 	sp "github.com/openebs/maya/pkg/sp/v1alpha1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes/fake"
 )
 
 var blockDeviceK8sClient *blockdevice.KubernetesClient
