+++ pkg/cstor/poolcluster/v1alpha1/build_test.go
@@ -18,7 +18,7 @@ package v1alpha1
 
 import (
 	apisv1alpha1 "github.com/openebs/maya/pkg/apis/openebs.io/v1alpha1"
-	poolspec "github.com/openebs/maya/pkg/cstorpoolcluster/v1alpha1/cstorpoolspecs"
+	poolspec "github.com/openebs/maya/pkg/cstor/poolcluster/v1alpha1/cstorpoolspecs"
 	"github.com/pkg/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"reflect"
