+++ tests/exporter/exporter_test.go
@@ -20,8 +20,8 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	apis "github.com/openebs/maya/pkg/apis/openebs.io/v1alpha1"
-	csp "github.com/openebs/maya/pkg/cstorpool/v1alpha3"
-	cv "github.com/openebs/maya/pkg/cstorvolume/v1alpha1"
+	csp "github.com/openebs/maya/pkg/cstor/pool/v1alpha3"
+	cv "github.com/openebs/maya/pkg/cstor/volume/v1alpha1"
 	pvc "github.com/openebs/maya/pkg/kubernetes/persistentvolumeclaim/v1alpha1"
 	sc "github.com/openebs/maya/pkg/kubernetes/storageclass/v1alpha1"
 	spc "github.com/openebs/maya/pkg/storagepoolclaim/v1alpha1"
