+++ tests/sts/sts_test.go
@@ -18,8 +18,8 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	k8s "github.com/openebs/maya/pkg/client/k8s/v1alpha1"
-	csp "github.com/openebs/maya/pkg/cstorpool/v1alpha2"
-	cvr "github.com/openebs/maya/pkg/cstorvolumereplica/v1alpha1"
+	csp "github.com/openebs/maya/pkg/cstor/pool/v1alpha2"
+	cvr "github.com/openebs/maya/pkg/cstor/volumereplica/v1alpha1"
 	pvc "github.com/openebs/maya/pkg/kubernetes/persistentvolumeclaim/v1alpha1"
 	pod "github.com/openebs/maya/pkg/kubernetes/pod/v1alpha1"
 	"github.com/openebs/maya/tests/artifacts"
